+++ Model/Storage/Storage.cs
@@ -7,14 +7,14 @@ namespace InventoryControl.Model
     public class Storage : IEntity, INamed, ITransferSpot
     {
         public static JoinTable ProductsNumberTable { get; }
-            = new JoinTable("ProductsNumber", typeof(Product), typeof(Storage), SqlType.INTEGER);
+            = new JoinTable("ProductsNumber", typeof(Product), typeof(Storage), SqlType.INT);
         public static Table<Storage> Table { get; } = new Table<Storage>
         (
             new List<Storage>() { new Storage(0, "Стандартный склад", "", 0) },
             new Column<Storage>("Name", SqlType.TEXT, (x) => x.Name,
                 Constraint.NotNull),
             new Column<Storage>("Address", SqlType.TEXT, (x) => x.Address),
-            new Column<Storage>("OwnerId", SqlType.INTEGER, (x) => x.ownerId,
+            new Column<Storage>("OwnerId", SqlType.LONG, (x) => x.ownerId,
                 Constraint.NotNull | Constraint.ForeighnKey("Counterparty"))
         );
 
@@ -29,7 +29,7 @@ namespace InventoryControl.Model
             set => ownerId = value.Id;
         }
         
-        public Storage(int id, string name, string address, int ownerId)
+        public Storage(long id, string name, string address, long ownerId)
         {
             this.Id = id;
             this.Name = name;
